pyproject.toml
[project]
name = "vox-timeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "qwen-tts>=0.0.5",
  "soundfile>=0.13.1",
  "torch>=2.9.1",
  "torchvision>=0.24.1",
  "openai-whisper>=20231117",
  "moviepy>=1.0.3",
  "pydub>=0.25.1",
  "gradio>=4.0.0",
  "transformers>=4.57.3",
  "accelerate>=1.12.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
